fix stuff and add a few comments
I did not have a look at the requirements for submitting a tutorial but pls have a look at Edzer's tutorial submission in 2017:
https://www.user2017.brussels/uploads/Edzer_tutorial_proposal_template.html
It specifically addresses the contents, the audience and learning objectives all of which is only marginally mentioned in our abstract. Furthermore, I think adding that you are the authors of Geocomputation with R might be a good idea for promoting the tutorial. 
Sorry for being a pain still I hope my suggestions help.